+++ spec/helper/MockHelper.ts
@@ -50,6 +50,7 @@ export namespace MockHelper {
       onSeeked: getEventDispatcherMock(),
       onRelease: getEventDispatcherMock(),
       onComponentViewModeChanged: getEventDispatcherMock(),
+      uiWrapperElement: generateDOMMock(),
     }));
 
     return new UiInstanceManagerMockClass();
@@ -60,6 +61,7 @@ export namespace MockHelper {
       addClass: jest.fn(),
       removeClass: jest.fn(),
       on: jest.fn(),
+      off: jest.fn(),
       html: jest.fn(),
       css: jest.fn(),
       width: jest.fn(),